[py3] Add a str-compatible MemoryMap implementation
I really hate to do this, but the more I dig into making a driver actually
work, I think this needs to be the default. I retained the bytes-clean
version as a base, and we can/should make new drivers use it, as well as
convert drivers to use it in this process. Bitwise still uses it under the
covers, so this is really just a usability shim to buy us some more time.

#495