[py3] Fix integer divison in from_NHz() helpers
NOTE: These appear to be unused, except for one use of from_GHz() in the
icq7 driver. AFAICT, these have always been wrong in the tree, so this
just converts the division and adds (wrong) tests to validate the division
change, but they should be fixed (along with the icq7 driver).

#495