[vx8] #1615 NFM support added
There is a 5+ years old issue which I could reproduce: the Yaesu VX-8
didn't support Chirp NFM (as in "half channel bandwidth", not as in
Yaesu's meaning: Yaesu calls the normal 25 kHz bandwidth "NFM", as
opposed to broadcast "WFM"; to achieve what's called "NFM" in Chirp,
one needs to enable "half deviation" in the VX-8).

This patch introduces Chirp NFM support to the VX-8 driver.
Tested on a VX-8DR.


73
Bernhard AE6YN

#1615